fix(trusty-mpm): surface the skills the every-run deploy declines (#4873)
Skill deployment already ran on all three run paths. Each reaches
`core::managed_config::ensure_managed_config_dir` through
`runtime::claude_code::prepare_managed_config`: `spawn` (fresh),
`spawn_resume` (`resume_managed`), and `build_inplace_resume_command`
(bare `tm` in a managed pane). Measured 2026-08-05 on the live machine,
49 of 52 skills under `$CLAUDE_CONFIG_DIR/skills/` were byte-identical
to the running binary's bundled source.

What was missing is any sign of it. `deploy_one_file` declines a
checksum-frozen or unmanaged skill and reports it only in the returned
`DeployStats`, which this caller discarded — so the three genuinely
frozen skills stayed stale on every run with no warning anywhere, and
read as "skill deployment never runs on resume". `ensure_managed_config_dir`
now emits one bounded warning naming them and pointing at
`tm doctor --fix-skills --include-frozen`, mirroring the agent-side
`deploy_summary_lines` that PR #4848 added for the same half of #4840.

The skip itself is unchanged and still correct: a hand-edited skill and
an unmanaged project-custom skill are both preserved, and an `mcp`-named
skill still never deploys.

Also corrects the `resume_managed` and `run_inplace_relaunch` comments
that claimed agent/skill redeploy and MCP injection were not re-run on
those paths. That claim was two-thirds false and is what routed this
defect report at the wrong function.

Tests move to `managed_config_tests.rs` (test-file SLOC cap) because
`managed_config.rs` was already at 416 of its 500-line production cap.

Closes #4873
